lock > mint > burn > unlock

+++ src/interfaces/IChildERC20.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.22;
 interface IChildERC20 {
     function bridge() external view returns (address);
 
-    function parent() external view returns (address token, uint64 chainId);
+    function parent() external view returns (address token, uint16 chainId);
 
     function mint(address to, uint amount) external;
